Fix props and states
